Checks the commit history of pull requests in CI.
The script runs on Travis and only checks the commit history when the commit is
from a pull request. It fails the build if the pull request contains merge
commits or does not start from the branch that it is being merged into. If there
is an unexpected exception, the script does not fail the build.