Only load one stats class with reflection.
Previously, the 'Stats' class used reflection to instantiate subclasses for both
'StatsManager' and 'StatsContextFactory'.  That design forced the 'StatsManager'
and 'StatsContextFactory' subclasses to only interact through their public APIs.

This commit changes 'Stats' so that it only instantiates 'StatsManager'.  The
'StatsManager' instance then creates the one 'StatsContextFactory' instance by
calling a constructor.  The main difference is that 'StatsManager' now knows
what subclass of 'StatsContextFactory' is in use, and all of the stats
implementation classes can interact through package-private methods.