Adds a class to represent a MeasurementDescriptor name.
We need to associate data with MeasurementDescriptors in several places in the
stats implementation.  For example, MeasurementMap maps MeasurementDescriptors
to recorded values.  Using MeasurementDescriptor as the key is not ideal,
because it is not obvious how the keys should be compared.  The keys could be
compared by name, since names are unique, or they could be compared by all
fields.

I think that we should always compare MeasurementDescriptors by name.  It is
more efficient, and it means that two MeasurementDescriptors with slightly
different descriptions won't be treated as distinct.  Having a separate key type
will make it clearer how the MeasurementDescriptors will be compared.

This commit just adds the class and doesn't change any existing APIs.