Split Distribution into mutable and immutable classes.
Distribution is an immutable AutoValue class, and MutableDistribution is a
mutable version of Distribution.  We can use MutableDistribution to store stats
as they are recorded and then copy them into a Distribution before returning
them in a View.  We can also use Distribution to store the distribution data in
DistributionAggregation.