Start adding a "tag" package.
This commit adds TagKey, TagValue, and TagSet to the new package. It does not
modify any classes in "stats", for backwards compatibility.  TagKey and TagValue
are the same as the corresponding classes in "stats", except that they use the
toString() methods generated by AutoValue.  TagSet is like StatsContext, except
that it uses AutoValue and doesn't contain any methods for recording stats or
serializing/deserializing.