Start adding a "tags" package.
This commit adds TagKey, TagChange, TagOp, and TagContext to the new package.
It does not modify any classes in "stats", for backwards compatibility.  TagKey
can be paired with values of type String, long, or boolean.  TagContext is
similar to StatsContext, except that it uses AutoValue and doesn't contain any
methods for recording stats or serializing/deserializing. It uses the TagChange
and TagOp classes to modify the tag values.