Implement StatsContext.record(...) using StatsManagerImplBase.record(...).
This change requires StatsContextImpl to store a reference to the
StatsManagerImplBase that is used to record the stats. It is a temporary change
that is only needed until we add the new tagging API.

This commit also updates the StatsContext tests so that they only record stats
for the one supported view.