Add a factory class for TagSet.
TagSetFactory will give us more flexibility to change the implementation of
TagSet in the future, or customize it based on the stats implementation.

This commit uses the factory to add mockable logging to TagSet.  Currently, the
logger is only used by TagSet.insert(...) to report when the key being inserted
already exists.

There are also a few miscellaneous improvements.