Replace StatsManager with StatsComponent class.  (Fixes #376)
StatsComponent is the class that is instantiated with reflection, to determine
which stats library implementation to use.  It contains a ViewManager,
StatsRecorder, and StatsContextFactory.  (The StatsContextFactory will be
removed when we replace StatsContext with TagContext.)

Breaking up the stats library functionality into multiple classes allows client
code to pass around only the parts that it needs.  Then when users mock the
stats library, they only need to mock the parts that their code uses.  For
example, a user can write unit tests for code that should record specific stats
by mocking StatsRecorder and asserting that StatsRecorder.record was called
with the correct arguments.

This change also makes the stats package more consistent with the trace package,
which has a TraceComponent that holds all other objects that provide tracing
functionality.