Move stats Context key to io.opencensus.tags.unsafe package.
This commit starts refactoring the stats io.grpc.Context key so that it can be
used to access a TagContext.  It also organizes the classes that interact with
the Context to be more similar to the tracing package.

The commit renames the constant from STATS_CONTEXT_KEY to TAG_CONTEXT_KEY and
moves it from io.opencensus.stats.ContextUtils to
io.opencensus.tags.unsafe.ContextUtils.  It also renames the
io.opencensus.stats.ContextUtils class to CurrentTagsUtils and makes it
package-private, since the only public member was the Context key.  The Context
key still has type Context.Key<StatsContext>, for now.