Add Tags and TagsComponent classes to the 'tags' package.
TagsComponent is an abstract class that holds the implementations for different
parts of the 'tags' package.  The class currently has one field, which is a
skeleton for a new TagContextFactory class.  It allows the implementation
library to determine the implementation of TagContext, which could be important
for performance.

This commit also adds a Tags class, which instantiates the TagsComponent with
reflection.  The design is very similar to the 'stats' package.  Tags
corresponds to the Stats class, and TagsComponent corresponds to StatsComponent.