Add a TagValue class.
This commit adds TagValue, which can represent a String, long, or boolean tag
value.  It replaces Object as the type for representing tag values in ViewData.

TagValue is an abstract class with three subclasses, similar to TagKey and Tag.
It provides a 'match' method for accessing the contained value. TagValueString
is now a subclass of TagValue instead of an independent, top-level class.