Clarifies the contract around scoping a trace context
Users may assume that the current trace context is the same as what they
placed into scope. For example, they might wrap a trace context and be
suprised that whatever they decorated it with is lost. This clarifies
the type-level javadoc on `TagContexts` to clarify this isn't supported.

See https://github.com/census-instrumentation/opencensus-java/pull/569#discussion_r136727376