Refactor TagContexts to make the current context methods final.
In order to make the methods final, this commit adds a non-final protected
method to TagContexts, transformTagContext, which handles any transformation of
the TagContext that must be done by the particular instance of TagContexts.  For
example, if the TagContexts allows tagging to be disabled, then
transformTagContext can return a no-op TagContext.