Move current TagContext logic from the API to the implementation (closes #632).
This commit makes all methods in Tagger and TagContextBuilder abstract.  It also
changes the no-op Tagger and TagContextBuilder so that they do not access the
current context.

This change gives the implementation more flexibility to control when it
accesses the current context.  For example, it could avoid setting the current
context when tagging is not enabled.  It also simplifies the code by not
splitting the current context logic between the API and implementation classes.