Make TagContext iterator method protected, and expose it through internal class.
This commit makes TagContext.unsafeGetIterator() protected and renames it to
iterator().  The protected modifier means that subclasses can specify the tags
that they contain, but users cannot access the tags of subclasses that they
did not create.  Since the stats and tagging implementations still need to
access the TagContext iterator, I added a class, io.opencensus.tags.Internal,
that contains a static method for accessing the iterator.

This change helps to separate the two uses of an iterator in TagContext:
giving the implementation access to tags in a TagContext subclass, which should
be exposed, and accessing any TagContext's tags, which should be treated as
internal.