Change ViewManager.getView to take a View.Name instead of a View.
Only one view can be registered per name, so only the name is needed to
unambiguously look up a registered view.

One benefit of this change is that users don't need to provide as much data.
This difference could be important when clients query views over the network.
Another benefit is that it eliminates the possibility that the input View's
non-name fields don't match the registered view.  One way to handle that case in
getView(View) is to check that the other fields match.  However, that is more
expensive than only looking up the name.