Add a separate case for OS X to .travis.yml (fixes #691).
This commit fixes a bug where TRAVIS_JDK_VERSION didn't match any cases on OS X,
because it wasn't set.  It probably wasn't set because the OS X image determines
the Java version, and the "jdk" field is ignored
(https://docs.travis-ci.com/user/reference/osx/#JDK-and-OS-X).

The OS X case avoids using jdk_switcher, because it isn't available on OS X.

This commit also adds a default case to fail the build when there is an unknown
OS or JDK.  The default case would have prevented the silent failure in #691.