Silently skip unknown TagContext fields during deserialization (issue #655).
Now TagContextBinarySerializerImpl.fromByteArray skips the rest of the output
when it encounters an unknown field, instead of throwing a
TagContextParseException.

This change will increase compatibility between the first version of the tagging
implementation and later versions, because the first version will be able to
accept TagContexts even when they contain newer fields.