Silently skip tags with unknown type during deserialization (issue #655).
Now TagContextBinarySerializerImpl.fromByteArray skips unknown tag types instead
of throwing a TagContextParseException.  When the tag has type long or boolean,
it only skips the individual tag.  When the tag has an unknown type, it skips
the rest of the input, since it doesn't know the tag's length.

This change will increase compatibility between the first version of the tagging
implementation and later versions, because the first version will be able to
accept tags that it supports, even if they are mixed with tags that it doesn't
support.