Fix the wrong use of notify and wait.
The previous code was using low-level synchronization based on
notify and wait - and it was wrong. The code allowed for a race condition
where notify is called before wait. In that case, notify will have no effect,
and wait will wait forever until interrupted by the test timeout.
I believe this to be the root cause of #793 (and the previous occurrences
in #540).

Just avoid the low-level synchronization and use Semaphore instead.