Run docker version to find whether Jaeger integration test can be performed.
Running `docker -v`, as done before this change, succeeds even when the Docker daemon is down, which means integration tests are attempted even though we cannot start Jaeger's container.
Running `docker version`, as done after this change, checks both the version of the CLI and the daemon, which fails when the daemon is down, and is what we really want here.