Replace Guava VisibleForTesting in opencensus-api (issue #1081).
This commit adds two annotations to replace the uses of Guava's
VisibleForTesting annotation in opencensus-api.  Each annotation has a more
specific meaning than VisibleForTesting, to make the uses of the annotations
more informative.

DefaultVisibilityForTesting - This annotation replaces the main use of
@VisibleForTesting, indicating that an element was changed from private to
package-private for testing.

PublicForTesting - This annotation is temporary and should be removed as part
of issue #977.