Set bucket bounds as le labels for Prometheus exporter.
Previously we don't set the bucket bounds for Prometheus Samples,
since bucket bounds is a built-in feature for Prometheus Histogram.
However, as #1164 suggested, we can manually set the bucket bounds
with the same label le that the built-in Histogram uses,
to get the same output as the built-in Prometheus Histogram Samples.