Allow overriding the Project ID used by OpenCensusTraceLoggingEnhancer.
The project ID used for formatting trace IDs can be overridden with the property
"io.opencensus.contrib.logcorrelation.stackdriver.OpenCensusTraceLoggingEnhancer.projectId".
The property can be set as a java.util.logging property or a system property.
OpenCensusTraceLoggingEnhancer continues to use the google-cloud-logging default
project ID when the projectId property is not set.

This commit also caches the project ID on the OpenCensusTraceLoggingEnhancer
instance instead of looking it up for every log entry.