Improve implementation of OpenCensusTraceContextDataInjector. (#1413)
This commit improves the thread safety of OpenCensusTraceContextDataInjector by
following the thread safety requirements in the Javadocs of the overridden
methods from ContextDataInjector.  It also handles the possibility of
ThreadContext.getThreadContextMap() returning null by adding a Nullable
annotation in a Checker Framework stub file and adding null checks to the code.