Temporarily remove "traceSampled" label from Stackdriver log correlation. (#1532)
Currently, opencensus-contrib-log-correlation-stackdriver adds the sampling
decision to a LogEntry as a custom label.  Once google-cloud-logging adds a
sampling decision field to its LogEntry class, we can set that field instead, so
that the value will be intepreted by Stackdriver.  However, removing the custom
label will be a breaking change.  This commit temporarily removes the custom
label so that we can make a non-experimental release of
opencensus-contrib-log-correlation-stackdriver and add the sampling decision to
the LogEntry later as a backwards-compatible change.