Bump jaeger trace exporter version
Using the jaeger trace exporter we started seeing `WARNING: A connection to http://jaeger:8080/ was leaked. Did you forget to close a response body? To see where this was allocated, set the OkHttpClient logger level to FINE` in our application logs.

I traced this back to jaeger client release `0.33.0` that fixes this issue with closing the connection.

The newer version of the jaeger client uses a new package name `io.jaegertracing`

https://github.com/jaegertracing/jaeger-client-java/pull/571
https://github.com/jaegertracing/jaeger-client-java/releases