Separate "tracestate" parsing, emit single format
This commit removes "tracestate" parsing from the `B3Format` class. If
we want that functionality in the future, a new `TracestateFormat` can
be introduced that delegates to various `TextFormat` classes (including
`B3Format`).

In addition, a new parameter is added to control whether fields are
injected using the multi-header format or the single-header format. Note
that neither `PropagationComponent` nor any upstream clients have yet
been modified to take advantage of the new parameter.

Addresses: #1400