Provider finer-grained control over B3 injection
In response to code review feedback, this commit allows clients to
specify single injection, multi injection, or both. The default is multi
injection. Although only single and multi are supported right now,
future variants can be supported. For example, if parent span is exposed
in the future, formats can be added that explicitly include or exclude
it. Note, support for "span kind"-specific injection formats is not
currently supported as that information is not available to
`TextFormat`.

Addresses: 1400